Move elasticsearch-pekko to elasticsearch
