+++ build.sbt
@@ -91,7 +91,7 @@ lazy val core = module(project, "core")
     )
   )
 
-lazy val elasticsearchPekko = module(project, "elasticsearch-pekko")
+lazy val elasticsearch = module(project, "elasticsearch")
   .dependsOn(testkit % Test)
   .settings(
     libraryDependencies ++= Seq(
@@ -183,7 +183,7 @@ lazy val apso = (project in file("."))
     circe,
     collections,
     core,
-    elasticsearchPekko,
+    elasticsearch,
     encryption,
     hashing,
     io,
@@ -198,7 +198,7 @@ lazy val apso = (project in file("."))
     circe,
     collections,
     core,
-    elasticsearchPekko,
+    elasticsearch,
     encryption,
     hashing,
     io,
