+++ elasticsearch/src/main/resources/reference.conf
