+++ elasticsearch/src/main/scala/com/kevel/apso/elasticsearch/config/Elasticsearch.scala
@@ -1,8 +1,8 @@
-package com.kevel.apso.elasticsearch.pekko.config
+package com.kevel.apso.elasticsearch.config
 
 import scala.concurrent.duration._
 
-import com.kevel.apso.elasticsearch.pekko.config.Elasticsearch.BulkInserter
+import com.kevel.apso.elasticsearch.config.Elasticsearch.BulkInserter
 
 case class Elasticsearch(
     host: String,
